requirements.txt
arrow==1.2.1
asgiref==3.4.1
attrs==25.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blessed==1.19.0
boto3==1.21.20
botocore==1.24.20
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
channels==4.2.2
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.8.0
django-cuser==2017.3.16
django-debug-toolbar==3.2.2
django-environ==0.11.2
django-extensions==3.1.3
django-filter==2.4.0
django-ipware==4.0.0
django-oauth-toolkit==2.1.0
django-picklefield==3.0.1
django-q==1.3.9
django-redis==5.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-swagger==2.1.2
django-storages==1.13.1
django-webpack-loader==3.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
ecdsa==0.18.0
fcm-django==1.0.5
firebase-admin==5.1.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
graphviz==0.20.1
grpcio==1.42.0
grpcio-status==1.42.0
hashids==1.3.1
httplib2==0.20.2
idna==3.3
importlib-resources==6.4.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.3.1
MarkupSafe==2.0.1
msgpack==1.0.2
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==8.3.2
pkgutil-resolve-name==1.3.10
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
redis==3.5.3
referencing==0.35.1
requests==2.26.0
requests-oauthlib==1.3.1
rpds-py==0.20.1
rsa==4.7.2
s3transfer==0.5.2
sentry-sdk==2.23.1
simplejson==3.17.5
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.12.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.20.2
